fetchAPI avec url generic pour avoir les 3 urls de l'API
