+++ src/App.tsx
@@ -4,15 +4,21 @@ import { Button } from "../src/ui/button/Button";
 import { Header } from "./ui/Header/Header";
 import { Banner } from "./ui/banner/Banner";
 import { Carousel } from "./ui/carousel/Carousel";
-import { Fetch } from "./ui/Fetch/fetch";
+import { Fetch } from "./ui/Fetch/FetchAPI";
 
 export const App = () => {
 
   // const [joke, setJoke] = useState<string | undefined>(undefined);
 
-  const Fetcher = Fetch(); // appelle ma const Fetch de Fetch.tsx
+  const FetcherPopular = Fetch('https://api.themoviedb.org/3/movie/popular?language=en-US&page=1') ; // appelle ma const Fetch de Fetch.tsx
+
+  console.log(FetcherPopular);
 
   
+  
+  
+
+
 
   return (
     <>
